refactor: post요청 body로 데이터 받도록 변경
- post요청 시 body로 데이터 받도록 변경
- 에러 처리
