+++ src/main/java/com/ssafy/solvedpick/image/dto/ImageSaveRequest.java
@@ -1,0 +1,14 @@
+package com.ssafy.solvedpick.image.dto;
+
+import lombok.*;
+
+@Getter
+@Builder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor(access = AccessLevel.PROTECTED)
+public class ImageSaveRequest {
+
+    private String originalFilename;
+    private String storedFilename;
+
+}
